Revert "Merge pull request #47061 from slava77/CMSSW_14_1_0_pre7/stripsTiming"
This reverts commit d288c81b480435b414132de7ccc4fe5bb958785e, reversing
changes made to 6eea5fa05e2eeb002918482c8ee4611c7349f861.
